# swaggerDocument.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Hỗ trợ du lịch",
        "version": "1.0.0",
        "title": "Smartourist App"
    },
    "schemes": {
        "http": "http",
        "https": "https"
    },
    "securityDefinitions": {
        "bearerAuth": {
            "name": "Authorization",
            "in": "header",
            "type": "apiKey",
            "description": "Bearer {token}"
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/user/deleteUser": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Xóa người dùng",
                "description": "Chỉ admin mới có thể xóa người dùng",
                "operationId": "deleteUser",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "uid",
                        "in": "query",
                        "description": "Mã uid của user",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Invalid UID supplied"
                    },
                    "404": {
                        "description": "USER NOT FOUND"
                    }
                }
            }
        },
...
    },
...
}